Updates after review 17.06
